# nodejs_space/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/attendance/bulk": {
      "post": {
        "operationId": "AttendanceController_bulkSave",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkAttendanceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Attendance"
        ]
      }
    },
    "/api/attendance/overview": {
      "get": {
        "operationId": "AttendanceController_getOverview",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceOverviewResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Attendance"
        ]
      }
    },
...
    "/api/progress": {
      "post": {
        "operationId": "ProgressController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProgressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Progress"
        ]
      },
      "get": {
        "operationId": "ProgressController_findByStudent",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
...
    "/api/academic-years": {
      "get": {
        "operationId": "AcademicYearsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademicYearListItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Academic Years"
        ]
      },
      "post": {
        "operationId": "AcademicYearsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcademicYearDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYearResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Academic Years"
        ]
      }
    },
    "/api/academic-years/{id}": {
      "patch": {
        "operationId": "AcademicYearsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAcademicYearDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYearResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Academic Years"
        ]
      }
    },
...
    "/api/report-cards/generate": {
      "post": {
        "operationId": "ReportCardsController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportCardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportCardResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Report Cards"
        ]
      }
    },
    "/api/report-cards": {
      "get": {
        "operationId": "ReportCardsController_findAll",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportCardResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Report Cards"
        ]
      }
    },
    "/api/report-cards/{id}/download": {
      "get": {
        "operationId": "ReportCardsController_getDownload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUrlResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Report Cards"
        ]
      }
    },
    "/api/events": {
      "get": {
        "operationId": "EventsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventListItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      },
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}/groups": {
      "post": {
        "operationId": "EventsController_createGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventGroupDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/groups/{groupId}": {
      "delete": {
        "operationId": "EventsController_removeGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/{id}/participants": {
      "post": {
        "operationId": "EventsController_addParticipant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddParticipantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/api/events/participants/{participantId}": {
      "delete": {
        "operationId": "EventsController_removeParticipant",
        "parameters": [
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
...
  },
  "info": {
    "title": "Api",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "FileUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "StudentAttendanceEntry": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "studentId",
          "status"
        ]
      },
      "TeacherAttendanceEntry": {
        "type": "object",
        "properties": {
          "teacherId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "teacherId",
          "status"
        ]
      },
      "BulkAttendanceDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "studentAttendance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentAttendanceEntry"
            }
          },
          "teacherAttendance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeacherAttendanceEntry"
            }
          }
        },
        "required": [
          "sessionId",
          "studentAttendance"
        ]
      },
      "BulkAttendanceResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "savedCount": {
            "type": "number"
          }
        },
        "required": [
          "success",
          "savedCount"
        ]
      },
      "SessionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "attendanceSubmitted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "date",
          "attendanceSubmitted"
        ]
      },
      "ClassSummaryDto": {
        "type": "object",
        "properties": {
          "totalSessions": {
            "type": "number"
          },
          "averageAttendance": {
            "type": "number"
          }
        },
        "required": [
          "totalSessions",
          "averageAttendance"
        ]
      },
      "StudentBreakdownDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "studentName": {
            "type": "string"
          },
          "present": {
            "type": "number"
          },
          "absent": {
            "type": "number"
          },
          "late": {
            "type": "number"
          },
          "excused": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "percentage": {
            "type": "number"
          }
        },
        "required": [
          "studentId",
          "studentName",
          "present",
          "absent",
          "late",
          "excused",
          "total",
          "percentage"
        ]
      },
      "AttendanceOverviewResponseDto": {
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionSummaryDto"
            }
          },
          "classSummary": {
            "$ref": "#/components/schemas/ClassSummaryDto"
          },
          "studentBreakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentBreakdownDto"
            }
          }
        },
        "required": [
          "sessions",
          "classSummary",
          "studentBreakdown"
        ]
      },
...
      "CreateProgressDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "progressMetricId": {
            "type": "string",
            "format": "uuid"
          },
          "classSessionId": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "type": "number"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "studentId",
          "progressMetricId",
          "classSessionId",
          "value"
        ]
      },
      "ProgressResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "progressMetricId": {
            "type": "string"
          },
          "classSessionId": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "studentId",
          "progressMetricId",
          "classSessionId",
          "value",
          "notes",
          "createdAt"
        ]
      },
      "ProgressEntryItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "date",
          "sessionId",
          "value",
          "notes"
        ]
      },
      "ProgressMetricListDto": {
        "type": "object",
        "properties": {
          "metricId": {
            "type": "string"
          },
          "metricName": {
            "type": "string"
          },
          "metricType": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgressEntryItemDto"
            }
          }
        },
        "required": [
          "metricId",
          "metricName",
          "metricType",
          "entries"
        ]
      },
      "ProgressListResponseDto": {
        "type": "object",
        "properties": {
          "metrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgressMetricListDto"
            }
          }
        },
        "required": [
          "metrics"
        ]
      },
...
      "AcademicYearListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "startDate",
          "endDate",
          "isActive",
          "createdAt"
        ]
      },
      "CreateAcademicYearDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "startDate",
          "endDate"
        ]
      },
      "AcademicYearResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "startDate",
          "endDate",
          "isActive"
        ]
      },
      "UpdateAcademicYearDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "GenerateReportCardDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "academicYearId": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          }
        },
        "required": [
          "studentId",
          "academicYearId"
        ]
      },
      "ReportCardResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "studentName": {
            "type": "string"
          },
          "academicYearName": {
            "type": "string"
          },
          "termName": {
            "type": "string",
            "nullable": true
          },
          "pdfFileId": {
            "type": "string"
          },
          "pdfUrl": {
            "type": "string"
          },
          "generatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "studentId",
          "studentName",
          "academicYearName",
          "termName",
          "pdfFileId",
          "pdfUrl",
          "generatedAt"
        ]
      },
      "EventListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "participantCount": {
            "type": "number"
          },
          "groupCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "date",
          "location",
          "participantCount",
          "groupCount"
        ]
      },
      "EventGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "EventDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventGroupDto"
            }
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventParticipantDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "date",
          "location",
          "groups",
          "participants"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "date"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        }
      },
      "CreateEventGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "AddParticipantDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}